refactor(libsinsp): split `sinsp_fdtable` and `sinsp_fdinfo`
