+++ userspace/libsinsp/fdinfo.h
@@ -24,9 +24,7 @@ limitations under the License.
 #include <libsinsp/state/table.h>
 
 #include <unordered_map>
-#include <vector>
 #include <memory>
-#include <functional>
 
 #ifdef _WIN32
 #define CANCELED_FD_NUMBER INT64_MAX
@@ -360,107 +358,4 @@ public:
 	uint64_t m_ino = 0;
 	int64_t m_pid = 0;  // only if fd is a pidfd
 	int64_t m_fd = -1;
-};
-
-/*@}*/
-
-// Forward declare sinsp_stats_v2 to avoid including metrics_collector.h here.
-struct sinsp_stats_v2;
-
-///////////////////////////////////////////////////////////////////////////////
-// fd info table
-///////////////////////////////////////////////////////////////////////////////
-class sinsp_fdtable : public libsinsp::state::built_in_table<int64_t> {
-public:
-	typedef std::function<bool(int64_t, sinsp_fdinfo&)> fdtable_visitor_t;
-
-	typedef std::function<bool(int64_t, const sinsp_fdinfo&)> fdtable_const_visitor_t;
-
-	sinsp_fdtable(sinsp* inspector);
-
-	inline std::unique_ptr<sinsp_fdinfo> new_fdinfo() const { return sinsp_fdinfo{}.clone(); }
-
-	sinsp_fdinfo* find(int64_t fd);
-
-	sinsp_fdinfo* add(int64_t fd, std::unique_ptr<sinsp_fdinfo> fdinfo);
-
-	inline bool const_loop(const fdtable_const_visitor_t callback) const {
-		for(auto it = m_table.begin(); it != m_table.end(); ++it) {
-			if(!callback(it->first, *it->second)) {
-				return false;
-			}
-		}
-		return true;
-	}
-
-	inline bool loop(const fdtable_visitor_t callback) {
-		for(auto it = m_table.begin(); it != m_table.end(); ++it) {
-			if(!callback(it->first, *it->second)) {
-				return false;
-			}
-		}
-		return true;
-	}
-
-	// If the key is present, returns true, otherwise returns false.
-	bool erase(int64_t fd);
-
-	void clear();
-
-	size_t size() const;
-
-	void reset_cache();
-
-	inline uint64_t get_tid() const { return m_tid; }
-
-	inline void set_tid(uint64_t v) { m_tid = v; }
-
-	// ---- libsinsp::state::table implementation ----
-
-	size_t entries_count() const override { return size(); }
-
-	void clear_entries() override { clear(); }
-
-	std::unique_ptr<libsinsp::state::table_entry> new_entry() const override;
-
-	bool foreach_entry(std::function<bool(libsinsp::state::table_entry& e)> pred) override {
-		return loop([&pred](int64_t i, sinsp_fdinfo& e) { return pred(e); });
-	}
-
-	std::shared_ptr<libsinsp::state::table_entry> get_entry(const int64_t& key) override;
-
-	std::shared_ptr<libsinsp::state::table_entry> add_entry(
-	        const int64_t& key,
-	        std::unique_ptr<libsinsp::state::table_entry> entry) override {
-		if(!entry) {
-			throw sinsp_exception("null entry added to fd table");
-		}
-		auto fdinfo = dynamic_cast<sinsp_fdinfo*>(entry.get());
-		if(!fdinfo) {
-			throw sinsp_exception("unknown entry type added to fd table");
-		}
-		entry.release();
-
-		return add_ref(key, std::unique_ptr<sinsp_fdinfo>(fdinfo));
-	}
-
-	bool erase_entry(const int64_t& key) override { return erase(key); }
-
-private:
-	sinsp* m_inspector;
-	std::unordered_map<int64_t, std::shared_ptr<sinsp_fdinfo>> m_table;
-	std::shared_ptr<sinsp_stats_v2> m_sinsp_stats_v2;
-
-	//
-	// Simple fd cache
-	//
-	int64_t m_last_accessed_fd;
-	std::shared_ptr<sinsp_fdinfo> m_last_accessed_fdinfo;
-	uint64_t m_tid;
-	std::shared_ptr<sinsp_fdinfo> m_nullptr_ret;  // needed for returning a reference
-
-private:
-	inline void lookup_device(sinsp_fdinfo& fdi) const;
-	const std::shared_ptr<sinsp_fdinfo>& find_ref(int64_t fd);
-	const std::shared_ptr<sinsp_fdinfo>& add_ref(int64_t fd, std::unique_ptr<sinsp_fdinfo> fdinfo);
 };
